Made separate module for hamburger and imported into navbar
