+++ next.config.js
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {}
+const nextConfig = {
+    images: {
+        domains: ['res.cloudinary.com'],
+    },
+}
 
 module.exports = nextConfig
